Alternate Testing
First part of alternative pixel-draw testing involving setting the pixels of a texture rather than instantiating a tile implemented, and seems to be working perfectly.
The next stage is to ensure the coordinates are being properly selected as they'll need to coincide with the coordinates of the int grid array.